Updated the appender to use RestSharp instead of WebRequest
Passed basic tests, still needs more work to be more confident in this version.
It's a work in progress.
