+++ src/log4stash.net35/packages.config
@@ -3,4 +3,5 @@
   <package id="ILRepack" version="2.0.13" targetFramework="net35" />
   <package id="log4net" version="2.0.8" targetFramework="net35" />
   <package id="Newtonsoft.Json" version="10.0.3" targetFramework="net35" />
+  <package id="RestSharp" version="105.2.3" targetFramework="net35" />
 </packages>
