+++ src/log4stash.net40/packages.config
@@ -2,5 +2,9 @@
 <packages>
   <package id="ILRepack" version="2.0.13" targetFramework="net40" />
   <package id="log4net" version="2.0.8" targetFramework="net40" />
+  <package id="Microsoft.Bcl" version="1.1.8" targetFramework="net40" />
+  <package id="Microsoft.Bcl.Async" version="1.0.168" targetFramework="net40" />
+  <package id="Microsoft.Bcl.Build" version="1.0.14" targetFramework="net40" />
   <package id="Newtonsoft.Json" version="10.0.3" targetFramework="net40" />
+  <package id="RestSharp" version="105.2.3" targetFramework="net40" />
 </packages>
