+++ src/log4stash/Authentication/BasicAuthenticationMethod.cs
@@ -1,9 +1,11 @@
 ﻿using System;
 using System.Text;
+using RestSharp;
+using RestSharp.Authenticators;
 
 namespace log4stash.Authentication
 {
-    public class BasicAuthenticationMethod : IAuthenticationMethod
+    public class BasicAuthenticationMethod : IAuthenticationMethod, IAuthenticator
     {
         public string Password { get; set; }
 
@@ -15,6 +17,14 @@ namespace log4stash.Authentication
             var encodedAuthInfo = Convert.ToBase64String(Encoding.ASCII.GetBytes(authInfo));
             var authorizationHeaderValue = string.Format("{0} {1}", "Basic", encodedAuthInfo);
             return authorizationHeaderValue;
+        }
+
+        public void Authenticate(IRestClient client, IRestRequest request)
+        {
+            var authInfo = string.Format("{0}:{1}", Username, Password);
+            var encodedAuthInfo = Convert.ToBase64String(Encoding.ASCII.GetBytes(authInfo));
+            var authorizationHeaderValue = string.Format("{0} {1}", "Basic", encodedAuthInfo);
+            request.AddHeader("Authorization", authorizationHeaderValue);
         }
     }
 }
