+++ src/log4stash/ElasticClient/IElasticsearchClient.cs
@@ -13,6 +13,6 @@ namespace log4stash
         AuthenticationMethodChooser AuthenticationMethod { get; set; }
         void PutTemplateRaw(string templateName, string rawBody);
         void IndexBulk(IEnumerable<InnerBulkOperation> bulk);
-        IAsyncResult IndexBulkAsync(IEnumerable<InnerBulkOperation> bulk);
+        bool IndexBulkAsync(IEnumerable<InnerBulkOperation> bulk);
     }
 }
