+++ src/log4stash/packages.config
@@ -3,4 +3,5 @@
   <package id="ILRepack" version="2.0.13" targetFramework="net45" />
   <package id="log4net" version="2.0.8" targetFramework="net45" />
   <package id="Newtonsoft.Json" version="10.0.3" targetFramework="net45" />
+  <package id="RestSharp" version="105.2.3" targetFramework="net45" />
 </packages>
